sgl-kernel/python/sgl_kernel/cutedsl_gemm.py
import cutlass.cute as cute
from cutlass.cute import *
import torch

@cute.kernel
def blockwise_gemm_kernel(
    A: Tensor, B: Tensor, C: Tensor,
    A_sf: Tensor, B_sf: Tensor,
    alpha: Tensor
):
    # Optimized blockwise GEMM using CuteDSL
    # Supports SM90/SM100 optimizations
    pass

@cute.jit
def cutedsl_fp8_blockwise_scaled_mm(
    mat_a: torch.Tensor,
    mat_b: torch.Tensor,
    scales_a: torch.Tensor,
    scales_b: torch.Tensor,
    out_dtype: torch.dtype
):
    # Host-side launch logic
    pass
