sgl-kernel/pyproject.toml
[build-system]
requires = [
  "scikit-build-core>=0.10",
  "torch>=2.8.0",
  "wheel",
]
build-backend = "scikit_build_core.build"

[project]
name = "sgl-kernel"
version = "0.3.20"
authors = [
  { name="Yineng Zhang", email="me@zhyncs.com" },
]
description = "Kernel Library for SGLang"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Environment :: GPU :: NVIDIA CUDA"
]
dependencies = [
  "nvidia-cutlass-dsl==4.4.2",
]

[project.urls]
"Homepage" = "https://github.com/sgl-project/sglang/tree/main/sgl-kernel"
"Bug Tracker" = "https://github.com/sgl-project/sglang/issues"

[tool.wheel]
exclude = [
  "dist*",
  "tests*",
]

[tool.scikit-build]
cmake.build-type = "Release"
minimum-version = "build-system.requires"

wheel.py-api = "cp310"
wheel.license-files = []
wheel.packages = ["python/sgl_kernel"]
